Use celerybeat to schedule tasks
This commit brings in changes to schedule celery tasks by using
celerybeat and add a task running every 10 minutes.

Set CELERYD_MAX_TASKS_PER_CHILD to 1 to avoid reactor restartable issued.

The Twisted reactor cannot be restarted, so once one spider finishes
running and crawler stops the reactor implicitly, that worker is useless.
The trick of setting the value to 1 here means to kill the worker which
ran your spider and replace it with a fresh one, which prevents the
reactor from being started and stopped more than once.

* to start celerybeat: celery -A nicetomeetyou beat
* to start worker: celery -A nicetomeetyou worker -l info